Arduino Sensor Code
Confirmed working code for temperature data and accelerometer

Temperature sensor data is completely done. Data is interpreted in a way such that the raw output data was converted into a readable Celcius reading. Work needs to be done to convert this into a signal to be send out by an arduino analog pin to be interpreted by the pi (the sensor reads between -55C and 125C, and ard's analog out resolution is 10 bits (0-1023), so work needs to be done to map that range analogously)

Accelerometer data is mostly done; some bugginess on how it works and the data is raw. However, since we only want tilt data, the data doesn't necessarily have to be converted into G's, but rather some trig needs to get done between the XYZ readings to calculate an angle.

Current sensor readings were done on Tuesday that indicate that it partially works, but the voltage must be stepped down because the sensor outputs between 0-12V, and the arduino can only handle 5V max. Once that is done, the voltage can simply be read VIA an analog pin